web/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
asgi-ipc==1.4.2
asgi-rabbitmq==0.5.4
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.23.0
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.4
beautifulsoup4==4.4.0
billiard==3.3.0.23
boto==2.40.0
cached-property==1.3.1
ccm==2.1.4
celery==3.1.25
certifi==2017.7.27.1
cffi==1.11.2
channels==1.1.8
characteristic==14.3.0
chardet==3.0.4
click==6.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
cryptography==2.1.2
daphne==1.3.0
decorator==4.1.2
diff-match-patch==20110725.1
Django==2.0.3
django-appconf==1.0.2
django-cart==1.0.4
django-celery==3.2.2
#django-celery-beat==1.1.1
django-channels==0.7.0
django-compressor==2.2
django-cors-headers==2.1.0
django-db-logger==0.1.6
django-filter==1.1.0
django-haystack==2.6.1
django-import-export==0.4.5
django-jenkins==0.110.0
django-jsonfield==1.0.1
django-paypal==0.2
django-picklefield==1.0.0
django-prettyjson==0.3.0
django-responsive==0.3.0
django-simple-captcha==0.5.5
django-tinymce==2.3.0
django-user-agents==0.3.2
django-wysiwyg==0.7.1
djangorestframework==3.7.3
djorm-ext-pgfulltext==0.10
easy-thumbnails==2.4.1
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
futures==2.2.0
geocoder==1.33.0
gunicorn==19.7.1
hashids==1.2.0
httplib2==0.10.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
instagram-scraper==1.5.5
ipaddress==1.0.18
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
jslint==0.6.0
jsonrpclib==0.1.7
kombu==3.0.37
Markdown==2.6.9
MarkupSafe==1.0
msgpack-python==0.4.8
mysqlclient==1.3.12
oauthlib==2.0.6
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
paramiko==2.3.1
pep8==1.7.1
pika==0.11.0
Pillow==4.3.0
posix-ipc==1.0.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycountry==17.5.14
pycparser==2.18
pyflakes==1.6.0
PyNaCl==1.1.2
pyOpenSSL==17.3.0
python-dateutil==2.5.2
python-memcached==1.58
pytz==2017.2
PyYAML==3.12
ratelim==0.1.6
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
restclient==0.11.0
rjsmin==1.0.12
service-identity==17.0.0
six==1.11.0
South==1.0.2
standardjson==0.3.1
tablib==0.12.1
timeutils==0.3.2
tqdm==4.19.4
Twisted==16.4.1
txaio==2.8.2
typing==3.6.2
ua-parser==0.7.3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
user-agents==1.1.0
vine==1.1.4
whois==0.7
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.4.3
woocommerce==1.2.1